Removing brainnet to avoid feature creep
