+++ cca_zoo/deepmodels/_dcca_base.py
@@ -26,7 +26,7 @@ class _DCCA_base(nn.Module):
         """
         raise NotImplementedError
 
-    def post_transform(self, *z_list, train=False) -> Iterable[np.ndarray]:
+    def post_transform(self, z_list, train=False) -> Iterable[np.ndarray]:
         """
         Some models require a final linear CCA after model training.
