+++ cca_zoo/deepmodels/dcca.py
@@ -55,7 +55,7 @@ class DCCA(_DCCA_base):
         z = self(*args)
         return self.objective.loss(*z)
 
-    def post_transform(self, *z_list, train=False):
+    def post_transform(self, z_list, train=False):
         if train:
             self.cca = MCCA(latent_dims=self.latent_dims)
             z_list = self.cca.fit_transform(z_list)
