+++ cca_zoo/deepmodels/dtcca.py
@@ -37,11 +37,10 @@ class DTCCA(DCCA):
             eps=eps,
         )
 
-    def post_transform(self, *z_list, train=False) -> Iterable[np.ndarray]:
+    def post_transform(self, z_list, train=False) -> Iterable[np.ndarray]:
         if train:
             self.cca = TCCA(latent_dims=self.latent_dims)
-            self.cca.fit(z_list)
-            z_list = self.cca.transform(z_list)
+            z_list = self.cca.fit_transform(z_list)
         else:
             z_list = self.cca.transform(z_list)
         return z_list
