+++ examples/plot_dvcca.py
@@ -19,7 +19,26 @@ from cca_zoo.deepmodels import (
     DCCAE,
     DVCCA,
 )
-from examples.utils import plot_reconstruction
+
+
+def plot_reconstruction(model, dataset, idx):
+    (x, y), _ = dataset[idx]
+    recon_x, recon_y = model.recon(x, y)
+    if isinstance(recon_x, list):
+        recon_x = recon_x[0]
+        recon_y = recon_y[0]
+    recon_x = recon_x.detach().numpy()
+    recon_y = recon_y.detach().numpy()
+    fig, ax = plt.subplots(ncols=4)
+    ax[0].set_title('Original View 1')
+    ax[1].set_title('Original View 2')
+    ax[2].set_title('Reconstruction View 1')
+    ax[3].set_title('Reconstruction View 2')
+    ax[0].imshow(x.detach().numpy().reshape((28, 28)))
+    ax[1].imshow(y.detach().numpy().reshape((28, 28)))
+    ax[2].imshow(recon_x.reshape((28, 28)))
+    ax[3].imshow(recon_y.reshape((28, 28)))
+
 
 n_train = 500
 n_val = 100
